+++ base-ts.js
@@ -1,5 +1,5 @@
 const merge = require('lodash.merge');
 
 module.exports = merge({},
-  require('./rules/ts/base'),
+  require('./rules/typescript'),
   require('./rules/common'));
