+++ rules/react.js
