+++ rules/typescript.js
